Move ipToCode to gradle
