+++ app/build.gradle
@@ -226,20 +226,8 @@ android {
     }
 }
 
-task ipToCode {
-    def inputFile = file("${projectDir}/geolite2_country_blocks_ipv4.csv")
-
-    def outputDir = "${buildDir}/generated/binary"
-    def outputFile = new File(outputDir, "geolite2_country_blocks_ipv4.bin")
-
-    outputs.file outputFile
-
-    doLast {
-        exec {
-            commandLine "kotlin", "ipToCode.kts", inputFile.absolutePath, outputFile.absolutePath
-            println "Generated binary file at: ${outputFile.absolutePath}"
-        }
-    }
+apply {
+    from("ipToCode.gradle.kts")
 }
 
 preBuild.dependsOn ipToCode
